clippy: workspace-wide auto-fix sweep (Phase A)
Applied clippy --fix with pedantic+nursery flags; fixed resulting
tree-sitter u32/usize type mismatch (node.child expects u32).

+++ crates/sloc-languages/src/lib.rs
@@ -2738,7 +2738,9 @@ pub mod ts {
         }
 
         for i in 0..node.child_count() {
-            if let Some(child) = node.child(i) {
+            #[allow(clippy::cast_possible_truncation)]
+            // child_count bounded by tree-sitter u32 capacity
+            if let Some(child) = node.child(i as u32) {
                 visit(child, ctx);
             }
         }
